+++ src/include/duckpgq/core/operator/iterative_length/iterative_length_task.hpp
@@ -26,12 +26,12 @@ private:
   double Explore(const std::vector<std::bitset<LANE_LIMIT>> &visit,
     std::vector<std::bitset<LANE_LIMIT>> &next,
     const std::vector<uint64_t> &v,
-    const std::vector<uint16_t> &e,
+    const std::vector<uint64_t> &e,
     size_t v_size, idx_t v_offset);
 
   void RunExplore(const std::vector<std::bitset<LANE_LIMIT>> &visit,
                 std::vector<std::bitset<LANE_LIMIT>> &next,
-                const std::vector<uint64_t> &v, const std::vector<uint16_t> &e, size_t v_size, idx_t v_offset);
+                const std::vector<uint64_t> &v, const std::vector<uint64_t> &e, size_t v_size, idx_t v_offset);
 private:
   ClientContext &context;
   shared_ptr<IterativeLengthState> &state;
